requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pdfplumber>=0.10.3
openai>=1.10.0
google-generativeai>=0.3.2
apscheduler>=3.10.4
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.1
aiofiles>=23.2.1
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
